OPS-105: (Part 1) New robust and unit-tested Certificate AutoSigning Backend.
This is part 1 of OPS-105, which is to move all of the code that will be
common for logging into RightScale, getting lists of servers, managing
config files, etc into a common module.

This also represents a massive rewrite of the autosign.rb script itself
-- moving the main code into an AutoSigner object that is instantiated
by the script at start-time. This object is able to handle far more
error scenarios and cleanly raise exceptions when certificates should
not be signed.

Logging has been greatly improved with the use of the 'logging' gem that
behaves more like the Python logging framework. Logs are outputted to
STDOUT always, but can be duplicated to a log file based on the 'debug'
setting in the config file.

Lastly, nearly full unit/spec tests for all of the methods have been
written. Appropriate mocks have been put in place so that no outbound
API calls are made, and real valid tests are now built using good and
bad CSR files.

OPS-105: (Part 2) New Hiera backend that leverages the RightScale object.

Using the Hiera pluggable backend model, this plugin leverages the
RightScale backend to supply unique values back into your Puppet
manifests.