Bugfix: Return nil if no servers are found.
We were accidentally returning an empty array if no
servers were found in RightScale matching our tag. An empty
array is a valid return, which stops the Hiera() lookup process
and never triggers it to fallback to the 'default' value
that a user configured in their lookup.

The correct behavior here is to return nil if the list
of values returned by RightScale translates into no servers found
matching a given tag prefix. This code patch fixes that
behavior and adds a proper unit test for it.